rename module to gaussian_cost.py
