+++ skchange/costs/gaussian_cost.py
